+++ Kernel/Modules/AgentLinkObject.pm
@@ -69,7 +69,7 @@ sub Run {
         $TemporarySourceTicketLink = 1;
     }
 
-   # do the permission check only if it is no temporary ticket link used while creating a new ticket
+    # do the permission check only if it is no temporary ticket link used while creating a new ticket
     if ( !$TemporarySourceTicketLink ) {
 
         # permission check
@@ -202,7 +202,7 @@ sub Run {
             }
         }
 
-        # output link block
+        # output link delete block
         $Self->{LayoutObject}->Block(
             Name => 'Delete',
             Data => {
@@ -210,6 +210,16 @@ sub Run {
                 SourceObjectNormal => $SourceObjectDescription{Normal},
             },
         );
+
+        # output special block for temporary links
+        # to close the popup without reloading the parent window
+        if ( $Form{Mode} eq 'Temporary' ) {
+
+            $Self->{LayoutObject}->Block(
+                Name => 'LinkDeleteTemporaryLink',
+                Data => {},
+            );
+        }
 
         # get already linked objects
         my $LinkListWithData = $Self->{LinkObject}->LinkListWithData(
@@ -453,7 +463,7 @@ sub Run {
         if ( $Form{Mode} eq 'Temporary' ) {
 
             $Self->{LayoutObject}->Block(
-                Name => 'TemporaryLink',
+                Name => 'LinkAddTemporaryLink',
                 Data => {},
             );
         }